Update index.md
Removed information which was inaccurate to legal standards, software released without licenses is all-rights reserved where the release of the software provided by Arcade1up did not contain a license they also do not own the software being distributed just a small portion of IP included that they have licenses to distribute. The rest of this software may cause the overall package to be deemed GPLv2, or Apache 2.0 licensing but the two also conflict with each-other. Ultimately it is possible the software distribution package not including Intellectual property despite being packaged by a 3rd party entity still falls within GPLv2 allowing it to be re-distributed.